translate-gem/requirements.txt
slack_bolt
python-dotenv
google-cloud-aiplatform
notion-client
notion-client
